requirements.txt
garth
python-dotenv
